+++ Assets/Scripts/Managers/Name Manager V2/NameManager.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public abstract class NameManager {
     abstract public string SetName(Drink drink);
-    abstract public int SetCost();
+    abstract public float SetCost();
 }
